tutor profile issue fixed

+++ lib/modules/instructor_profile/models/teacher_review_response.dart
@@ -52,7 +52,7 @@ class Reviews {
   String? ratedToRef;
   String? ratedTo;
   String? satisfaction;
-  double? rating;
+  int? rating;
   String? feedback;
   String? createdAt;
